requirements.txt
streamlit>=1.56.0
pandas>=2.2.3
numpy>=2.1.0
matplotlib>=3.9.2
seaborn>=0.13.2
plotly>=5.24.1
Pillow>=10.4.0
requests>=2.32.3
pytest>=8.3.3
